Refactor Board class creating Aliens class to group to group the behavior of aliens as a whole
