+++ src/main/java/org/fundacionjala/at15/spaceinvaders/Board.java
@@ -4,22 +4,17 @@ import java.awt.event.KeyEvent;
 import java.awt.event.KeyListener;
 import javax.swing.JPanel;
 import java.awt.*;
-import java.util.ArrayList;
-import java.util.List;
 import static org.fundacionjala.at15.spaceinvaders.Constants.Alien.*;
 import static org.fundacionjala.at15.spaceinvaders.Constants.Board.*;
 import static org.fundacionjala.at15.spaceinvaders.Constants.Block.*;
 
 public class Board extends JPanel {
-    private List<Alien> aliens;
-    private int alienDeltaX = ALIEN_DELTA_X;
+    private Aliens aliens = new Aliens(ALIEN_ROWS, ALIEN_COLUMNS);
     private Ship ship = new Ship();
     private Gun gun = new Gun(ship);
     private Block block = new Block(ASTEROID_LIFE);
 
     public Board() {
-        aliens = new ArrayList<>();
-        createAliens(ALIEN_ROWS, ALIEN_COLUMNS);
         addKeyListener(new KeyListener() {
             @Override
             public void keyTyped(KeyEvent e) {
@@ -39,20 +34,6 @@ public class Board extends JPanel {
         setFocusable(true);
     }
 
-    public List<Alien> getAliens() {
-        return this.aliens;
-    }
-
-    public void createAliens(int rows, int columns) {
-        for (int yIndex = 0; yIndex < rows; yIndex++) {
-            for (int xIndex = 0; xIndex < columns; xIndex++) {
-                Alien alien = new Alien(
-                        ALIEN_INIT_X + ALIEN_SEPARATION * xIndex,
-                        ALIEN_INIT_Y + ALIEN_SEPARATION * yIndex, ALIEN_WIDTH, ALIEN_HEIGHT);
-                this.aliens.add(alien);
-            }
-        }
-    }
     public static void pause() {
         try {
             Thread.sleep(SLEEP);
@@ -74,34 +55,12 @@ public class Board extends JPanel {
             gun.move();
         }
 
-        for (Alien alien : this.aliens) {
+        for (Alien alien : this.aliens.getAliens()) {
             alien.paint(g);
         }
         pause();
-        this.moveAliens();
+        this.aliens.moveAliens();
         g.dispose();
         repaint();
-    }
-
-    public void moveAliens() {
-        for (Alien alien : this.aliens) {
-            int posX = alien.getPosX();
-            if (posX == 0 && this.alienDeltaX == -ALIEN_DELTA_X) {
-                this.alienDeltaX = ALIEN_DELTA_X;
-
-                List<Alien> aliens2 = aliens;
-                for (Alien alien2 : aliens2) {
-                    alien2.setPosY(alien2.getPosY() + ALIEN_SEPARATION);
-                }
-            }
-            if (posX == BOARD_WIDTH - ALIEN_WIDTH && this.alienDeltaX == ALIEN_DELTA_X) {
-                this.alienDeltaX = -ALIEN_DELTA_X;
-                List<Alien> aliens2 = aliens;
-                for (Alien alien2 : aliens2) {
-                    alien2.setPosY(alien2.getPosY() + ALIEN_SEPARATION);
-                }
-            }
-            alien.moveX(this.alienDeltaX);
-        }
     }
 }
